moved coordinates and bounds_to_coordinates out of chips.py and into functions.py, as they are general grid based functions and not specifically for chips.  Generalized coordinates to read the grid name from its parameter so coordinates now supports tiles and chips.  Added test for coodinates to check both grids.
