+++ merlin/chips.py
@@ -14,77 +14,6 @@ import requests
 logger = logging.getLogger(__name__)
 
 
-def coordinates(ulx, uly, lrx, lry, grid, cfg):
-    """Returns all the chip coordinates that are needed to cover a supplied
-    bounding box.
-
-    Args:
-        ulx  (float): upper left x
-        uly  (float): upper left y
-        lrx  (float): lower right x
-        lry  (float): lower right y
-        grid (dict): {'sx': 3000, 'sy': 3000, 'rx': 1, 'ry': -1}
-        cfg  (dict): A Merlin configuration
-
-    Returns:
-        tuple: tuple of tuples of chip coordinates ((x1,y1), (x2,y1) ...)
-
-    This example assumes chip sizes of 500 pixels.
-
-    Example:
-        >>> chip_coordinates = coordinates(ulx=-1000, 
-                                           uly=1000, 
-                                           lrx=-500, 
-                                           lry=500,
-                                           grid={'sx': 500, 'sy': 500}, 
-                                           cfg={'snap_fn': some_func})
-
-        ((-1000, 500), (-500, 500), (-1000, -500), (-500, -500))
-    """
-    
-    # get the snap_fn from Merlin config
-    snap_fn = cfg.get('snap_fn')
-
-    # snap start/end x & y
-    start_x, start_y = get_in(['chip', 'proj-pt'], snap_fn(x=ulx, y=uly))
-    end_x,   end_y   = get_in(['chip', 'proj-pt'], snap_fn(x=lrx, y=lry))
-
-    # get x and y scale factors multiplied by reflection
-    x_interval = grid.get('sx') * grid.get('rx')
-    y_interval = grid.get('sy') * grid.get('ry')
-    
-    return tuple((x, y) for x in np.arange(start_x, end_x + x_interval, x_interval)
-                        for y in np.arange(start_y, end_y + y_interval, y_interval))
-
-
-def bounds_to_coordinates(bounds, grid, cfg):
-    """Returns chip coordinates from a sequence of bounds.  Performs minbox
-    operation on bounds, thus irregular geometries may be supplied.
-
-    Args:
-        bounds: a sequence of bounds.
-        grid (dict): {'sx': SCALE_FACTOR, 'sy': SCALE_FACTOR}
-        cfg (dict): {'snap_fn': some_func}
-
-    Returns:
-        tuple: chip coordinates
-
-    Example:
-        >>> xys = bounds_to_coordinates(
-                                    bounds=((112, 443), (112, 500), (100, 443)),
-                                    grid={'sx': 3000, 'sy': 3000}, 
-                                    cfg={'snap_fn': some_func})
-        >>> ((100, 500),)
-    """
-
-    return coordinates(ulx=f.minbox(bounds)['ulx'],
-                       uly=f.minbox(bounds)['uly'],
-                       lrx=f.minbox(bounds)['lrx'],
-                       lry=f.minbox(bounds)['lry'],
-                       grid=grid,
-                       cfg=cfg)
-
-
 def locations(x, y, cw, ch, rx, ry, sx, sy):
     """Computes locations for array elements that fall within the shape
     specified by chip_spec['data_shape'] using the x & y as the origin.
