+++ test/resources/chipmunk-ard-v1-cassette.yaml
@@ -2054,4 +2054,55 @@ interactions:
       Date: ['Thu, 28 Dec 2017 22:11:02 GMT']
       Server: [http-kit]
     status: {code: 200, message: OK}
+- request:
+    body: null
+    headers:
+      Accept: ['*/*']
+      Accept-Encoding: ['gzip, deflate']
+      Connection: [keep-alive]
+      User-Agent: [python-requests/2.18.1]
+    method: GET
+    uri: http://localhost:5656/grid
+  response:
+    body: {string: '[{"name":"tile","proj":null,"rx":1.0,"ry":-1.0,"sx":150000.0,"sy":150000.0,"tx":2565585.0,"ty":3314805.0},{"name":"chip","proj":null,"rx":1.0,"ry":-1.0,"sx":3000.0,"sy":3000.0,"tx":2565585.0,"ty":3314805.0}]'}
+    headers:
+      Content-Length: ['207']
+      Content-Type: [application/json; charset=utf-8]
+      Date: ['Fri, 16 Feb 2018 21:40:40 GMT']
+      Server: [http-kit]
+    status: {code: 200, message: OK}
+- request:
+    body: null
+    headers:
+      Accept: ['*/*']
+      Accept-Encoding: ['gzip, deflate']
+      Connection: [keep-alive]
+      User-Agent: [python-requests/2.18.1]
+    method: GET
+    uri: http://localhost:5656/grid/snap?x=0&y=0
+  response:
+    body: {string: '{"tile":{"proj-pt":[-15585.0,14805.0],"grid-pt":[17.0,22.0]},"chip":{"proj-pt":[-585.0,2805.0],"grid-pt":[855.0,1104.0]}}'}
+    headers:
+      Content-Length: ['121']
+      Content-Type: [application/json; charset=utf-8]
+      Date: ['Fri, 16 Feb 2018 21:40:40 GMT']
+      Server: [http-kit]
+    status: {code: 200, message: OK}
+- request:
+    body: null
+    headers:
+      Accept: ['*/*']
+      Accept-Encoding: ['gzip, deflate']
+      Connection: [keep-alive]
+      User-Agent: [python-requests/2.18.1]
+    method: GET
+    uri: http://localhost:5656/grid/snap?x=3000&y=-3000
+  response:
+    body: {string: '{"tile":{"proj-pt":[-15585.0,14805.0],"grid-pt":[17.0,22.0]},"chip":{"proj-pt":[2415.0,-195.0],"grid-pt":[856.0,1105.0]}}'}
+    headers:
+      Content-Length: ['121']
+      Content-Type: [application/json; charset=utf-8]
+      Date: ['Fri, 16 Feb 2018 21:40:40 GMT']
+      Server: [http-kit]
+    status: {code: 200, message: OK}
 version: 1
