+++ test/test_chips.py
@@ -14,32 +14,7 @@ from numpy.random import randint
 import numpy as np
 import test
 
-
-def chip_grid(config):
-    return first(filter(lambda x: x['name'] == 'chip', config.get('grid_fn')()))
-
-
-@test.vcr.use_cassette(test.cassette)
-def test_coordinates():
-
-    _cfg     = cfg.get('chipmunk-ard', env=test.env)
-    grid     = chip_grid(_cfg)
-    expected = ((-585.0, 2805.0), (-585.0, -195.0), (2415.0, 2805.0), (2415.0, -195.0))
-    result   = chips.coordinates(ulx=0, uly=0, lrx=3000, lry=-3000, grid=grid, cfg=_cfg)
-    assert expected == result
-
-
-@test.vcr.use_cassette(test.cassette)
-def test_bounds_to_coordinates():
-    _cfg     = cfg.get('chipmunk-ard', env=test.env)
-    grid     = chip_grid(_cfg)
-    expected = ((-3585.0, 5805.0), (-3585.0, 2805.0), (-585.0, 5805.0), (-585.0, 2805.0))
-    result   = chips.bounds_to_coordinates(bounds=((0, 0), (-590, 0), (0, 2806)),
-                                           grid=grid,
-                                           cfg=_cfg)
-    assert expected == result
-    
-
+   
 def test_locations():
     params = {'x': 0,
               'y': 0,
